#7 fix : envp, set을 모두 고려한 env substituter
